msm: kgsl: Add a KGSL_MEMFLAGS_VBO_NO_MAP_ZERO flag
Not all VBOs require the zero page mapped. These VBOs can use the
KGSL_MEMFLAGS_VBO_NO_MAP_ZERO flag to reduce time spent on allocate and
bind operations. VBOs that use this flag are mapped only to the ranges
bound to it. So, when binding and cleaning up those VBOs make sure we only
unmap the ranges that were already mapped.

Change-Id: I995db2d94f5e2c58ebdea3c6de1e4bc3df841093
Signed-off-by: Lynus Vaz <quic_lvaz@quicinc.com>