msm: kgsl: Query software status of hardware fences
If GMU detects that certain external hardware fences (part of a
sync object) haven't signaled in a given amount of time, GMU will
send F2H_MSG_SYNCOBJ_QUERY packet to query the software status of
these hardware fences. In response to this packet, kgsl will look
up these hardware fences (in the queried sync object) and indicate
to GMU whether the corresponding dma fences are signaled (or not)
via the same packet. In case a dma fence is signaled, add a log to
indicate the dma fence is signaled and the hardware fence isn't.

Change-Id: Iada8a8a2641cf74c591604b8236c8555d1fb4d2f
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>