msm: kgsl: Add support to handle hardware fence timeout
If GMU detects that a certain external hardware fence (part of a sync
object) is not signaled via the hardware path, GMU will query the
software status of that fence. Say, kgsl indicates to GMU that the
software status of the fence is signaled, GMU will wait for another
epoch. GMU will then send a F2H_MSG_CONTEXT_BAD packet if it still detects
that the hardware fence is not signaled via the hardware path. This
packet contains the context id (that was waiting for this external
hardware fence) and the sync object timestamp (to which the external
hardware fence belongs). Kgsl will log all the fences (and some metadata)
that are part of this sync object and trigger snapshot and recovery.

Change-Id: If7ece0d4e45b3ff4e0af05f9233752b99b0bb123
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>