msm: kgsl: Add support for mem alloc size alignment
The GMU VA mapping size of HFI_MEMKIND_MMIO_IPC_CORE mem alloc
request has a requirement that its VA size should be aligned to
power of two (as required by MPU). But, the actual size of this
mem alloc request is not necessarily going to be a power of two.
To honor the MPU requirement in this case, add support for a mem
alloc request to specify its size alignment requirements as a
power of two.

Change-Id: Ie62e12d1fb18922f812e7cbc55807e85ad79cec4
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>