msm: kgsl: Use a hint to improve gpuaddr allocation time
GPU VAs are allocated by looking for the first fit. In some allocation
patterns we might spend time looking for a hole. Instead, store a hint
based on the last allocated gpuaddr and use this to improve the search
time. Since this may result in VA space fragmentation do not use this hint
if limited to a 32-bit address space, or on global and secure pagetables.

Change-Id: Ib17bc300f53e2eaae578872a8665e5c36ae30fcf
Signed-off-by: Lynus Vaz <quic_lvaz@quicinc.com>