msm: kgsl: Modify tracking of inflight hardware fences
Currently, whether a hardware fence is sent to TxQueue or not is
tracked inline with each submission. Plus, all inflight hardware
fences from all contexts are tracked in a single list. This can
cause latency in the submission loop. Instead, keep track of them
in per context list and walk the list whenever kgsl receives a
TS_RETIRE packet from the GMU for a particular context.

Change-Id: If3c7179b21ef1bfdaddf4da011e8ab51111d1aa1
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>