mmrm: Round the client-set value for peak power calculations
MMRM clients may set a clock rate which is greater than max supported
rate for the clock. To perform peak power calculations, round the rate
with clk_round_rate() to get the rounded (max supported) value. Use the
client-set value to set the rate to clock driver, which internally
rounds the rate to the supported value also.

Change-Id: Ie9b4f5077244c8f43a2f7de2f9d6b284843091f1
Signed-off-by: Mihir Ganu <quic_mganu@quicinc.com>