msm: kgsl: Limit the number of unack'd hardware fence packets
Since we no longer wait for acks when sending hardware fences to GMU,
GMU can run out of its msg pool if there is a deluge of hardware fence
packets in a short span of time. This leads to GMU errors and reset/
recovery. To avoid running out of msgpool, make sure we only allow a
certain number of unacked hardware fence packets. When this limit is
reached, kgsl will put the userspace threads (that are requesting for
a dma fence) to sleep until the count of unacked hardware fences drops
to a desired threshold. However, it is possible that submission loop
wants to send a (already created) hardware fence to GMU, but the limit
is reached. We don't want to put this thread to sleep because it holds
dispatcher and device mutex. Instead, keep track of this context and
fence and handle it once the number of unacked hardware fence packets
drops to a desired threshold. Also, make sure we halt any further
cmdbatch submissions to the GPU until the number of unacked
hardware fence packets drops to a desired threshold. This makes sure
we don't end up with more than one deferred contexts. If there is a
concurrent fault, wake up any such sleeping threads as reset/recovery
takes a while to complete and we don't want to unnecessarily keep the
threads sleeping until then. Once reset/recovery is complete, handle
the deferred context (if there is one) and re-open the gates to allow
creation of new hardware fences.

Change-Id: I9ad2821aec26175fd18e79596d1319c07d9ff436
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>