msm: kgsl: Do not create hardware fences during fault recovery
Fault recovery usually takes a while to complete, which can cause
userspace to queue a substantial number of submissions and their
corresponding hardware fences. Post recovery, this can lead to a
deluge of hardware fence packets to GMU in a very short span of time.
To avoid this, do not create any new hardware fences until fault
recovery is complete.

Change-Id: I0f0a5d4ebf48a07e553b9744e67109b90eb8b892
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>