kgsl: Dump shader memory at the end of the snapshot sequence
Some gpu faults make the gpu shader processor run in an infinite
loop, which can cause the host or crash dumper access to shader
block to time out. This can result in incomplete data in snapshot.
Hence move the shader memory dump to the end of the sequence so that
we dump the majority of the gpu hw blocks information in the snapshot.

Change-Id: Ie343d0ed44b91ee273bcf378c02e4f1db5756ec1
Signed-off-by: Hareesh Gundu <quic_hareeshg@quicinc.com>