msm: kgsl: Handle sync object with software fence correctly
It is possible that even with hardware fences enabled, kgsl
will create a software only fence if the corresponding timestamp
has been retired by the GPU. Now, assume kgsl receives an input
fence array which has this (signaled) software only fence as one
of its children and say has another child fence which is a hardware
fence. Currently, kgsl will not send this sync object to GMU
because one of its children is a software fence. Whereas it would
be more appropriate to ignore the software fence (if it is signaled)
and send the rest of the hardware fences to GMU.

Change-Id: I3d22d689ff84308ec4026beedc2f7b0c75b01d16
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>