msm: kgsl: Update drawctxt internal_timestamp correctly
We have a corner case where we send a context unregister hfi packet
to GMU with timestamp 0 instead of the correct timestamp. This is
how it happens -
Say the last submission from a context is timestamp 10. Say, there is
a concurrent fault (from another context) before this timestamp is
retired. As part of reset sequence, the internal timestamp of this
context gets reset to 0. After GMU and GPU reset, we re-submit
timestamp 10 to the GMU. Say the context detaches at this point so
kgsl will send the context unregister packet with timestamp 0. This
causes a GMU assert because the timestamp should have been 10 instead.
Fix this by setting the internal timestamp in the target specific
submission function, which is common to both regular submission path
as well as replay path during reset/recovery.

Change-Id: I1049f08c9619929a8cd872fbcc24f51074752a6a
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>