mm-driver: Refactor to separate modules into packages
Currently, all of the modules are built from a single Bazel package.
This complicates integration with the vendor build because the
Android build system builds the modules individually.

Split the modules into their own Bazel packages to align more closely
with the Android build system's expectations and easy to hook bazel build.

Change-Id: I100e9ec9edbe96212089a5944cbba4d6677ff83a
Signed-off-by: Varsha Suresh <quic_varssure@quicinc.com>