msm: kgsl: Explicitly set size of hfi packets
This fixes an issue with gen7_hfi_queue_write() (and a few other
functions). These functions do not have any means to know/validate
the packet size which is supposed to be written to the specific
queue. Hence, explicitly pass the size of the packet to the
function and use that to set the size in the packet header.

Change-Id: I44da84e2d7bcea83fddab0e4f317aa0f5c766805
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>