ubwcp: disable range translation before flush
Ensure the buffer range translation is disabled before calling
the UBWC-P flush.
This is to avoid a potential HW hang due to
speculative lines being in the cache at time of DMA un-mapping.

We are still awaiting final confirmation on the programming
sequence. Once confirmed this code will likely need
re-factoring.

It is not clear if the isb() calls are required so we
may be able to optimize them out later.

Change-Id: Ib2e6d5d26cf9a47c9833b98c34d985484b58f46e
Signed-off-by: Liam Mark <quic_lmark@quicinc.com>
Signed-off-by: Linux Image Build Automation <quic_ibautomat@quicinc.com>