ubwcp: disable range translation before flush
Ensure the buffer range translation is disabled before calling
the UBWC-P flush.
This is to avoid a potential HW hang due to
speculative lines being in the cache at time of DMA un-mapping.

We are still awaiting final confirmation on the programming
sequence. Once confirmed this code will likely need
re-factoring.

It is not clear if the isb() calls are required so we
may be able to optimize them out later.

Change-Id: I65e9246f60bbbce8631a511ecf4a486298d93532
Signed-off-by: Liam Mark <quic_lmark@quicinc.com>