video: driver: fix the return type during open instance
There is a limit to the number of video instances which can be opened.
When the number of instances are exhausted, there is a possibility
that any ongoing instance is released. In such scenario, new instance
can be accepted. Hence keep the return type to suggest the client to
try opening the video instance again.

Change-Id: I2241a5458bbe55a2de1c58ad85440b2b920ae205
Signed-off-by: Vikash Garodia <quic_vgarodia@quicinc.com>