kgsl: gmu: Add support for GMU virtual register bank
There is a need for sharing certain bits of information
between kgsl and GMU before GMU has booted. So far, the
only way to do that is to use GMU scratch registers. But,
we only have a very limited number of scratch registers.
So, allocate a buffer in system memory such that each
dword in this buffer acts as a virtual scratch register.

Change-Id: Ib7a265283ec2cd04897a24cc974c6c2e4adebb11
Signed-off-by: Hareesh Gundu <quic_hareeshg@quicinc.com>