msm: kgsl: Loop through the msg queue for start msg
While sending start msg to GMU with multiple tasks running it is possible
that GMU expects response for mem alloc and perfcounter allocation from
two different tasks. Since KGSL clears the interrupt and read only one
message, it is possible to have unread messages in the msg queue, causing
mismatched responses and eventually timeout due to unprocessed messages.

When an interrupt is received, read the msg queue until it's empty before
polling for the next interrupt.

Change-Id: I41ae4d24b8fa0fed03fb3b662897f6f05ff20bce
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor <quic_mmandaya@quicinc.com>