kgsl: hwsched: Fix HFI sequence number wrap issue
When comparing the ack, make sure the entire header is compared instead
of just the sequence number. This is required because two packets
(waiting for their acks) can have the same sequence number (once it
wraps around). Use a different sequence number generator for cmdbatch
submissions to context queues or dispatch queues to reduce the chance
of wrapping around the cmdq sequence number. For the same reason, use a
different sequence number generator for hardware fence packets as well.
Remove instances where the sequence number is getting updated twice for
the same packet.

Change-Id: I56232a3b5cf74b725f9572bd34eb4041774dc6d1
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>