kgsl: hwsched: Clear hfi interrupt only if queue is not empty
While processing MSG_START packet, we clear the interrupt before reading
the msg queue irrespective of whether the queue has content in it. If
there is no content in the queue, the irq clear write could be posted right
after a new valid interrupt by GMU, causing us to timeout waiting for the
new interrupt. Clear the interrupt only if the queue has content in it.
Also add a memory barrier to explicitly make sure the write is posted.

Change-Id: I8c57a558cbe4b8e422f1e2e4d4a8dec936d36ec1
Signed-off-by: Mohammed Mirza Mandayappurath Manzoor <quic_mmandaya@quicinc.com>