kgsl: Clear GPU wake on touch flag on slumber entry
The wake on touch flag is only cleared when work is submitted to
the GPU after wakeup from a touch event.
If the GPU is woken up as a result of a write to the touch_wake
sysfs node and no work is submitted, this flag remains set
even though the GPU enters SLUMBER. This prevents the next touch
event from triggering early GPU wakeup. So, clear the flag when
GPU enters SLUMBER.

Change-Id: I5cf7fe4950602f96d87cfd646d73dc43e7db1ee2
Signed-off-by: Sushmita Susheelendra <quic_ssusheel@quicinc.com>
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>