BACKPORT: disp: msm: dsi: add DSI peripheral flush support
Adds argument to function used to send commands to send it using the
peripheral flush signal from MDP. Enables MDP to send the peripheral
flush signal to DSI to trigger command tx. This does not ensure that
MDP actually sends the signal, that needs to be done on a case by case
basis.

upstream-commit: https://git.codelinaro.org/clo/la/platform/vendor/opensource/display-drivers/-/commit/c4c4d5869cde36f8aefc3899db6f41873f0c773e

[inferno0230: Adapt for current codebase]

Change-Id: I272b53cea9fcadd72f876f4a71aeddbefb1ed40c
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>
Signed-off-by: inferno0230 <mail@inferno0230.in>